refactor: modularize server routes
